[6336] Move the retrieval of the connector tools in a dedicated hook
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6336
